pyproject.toml
[tool.poetry]
name = "recipesbook"
version = "0.5.0"
description = ""
authors = ["Sevenamber <sevenamber@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
django-environ = "^0.8.1"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
